removing unnecessary things and more
