+++ engine/src/main/java/org/destinationsol/game/SolGame.java
@@ -545,6 +545,4 @@ public class SolGame {
     public HullConfigManager getHullConfigManager() {
         return hullConfigManager;
     }
-
-    public boolean isInvincible(){return hero.isInvincible();}
 }
